+++ lib/plug_rest/exceptions.ex
@@ -1,21 +1,33 @@
-defmodule PlugRest.ResourceError do
+defmodule PlugRest.RuntimeError do
   @moduledoc """
-  Raised for Resource errors
-
-  The exception takes an atom describing the error status as detailed
-  in `Plug.Conn.Status`, and an optional error message.
+  Raised for Runtime errors
   """
 
   defexception [:message, plug_status: 500]
+end
 
-  def exception(args) when is_list(args) do
-    # Use `keyfind` to work around type error when using Keyword on args
-    {:status, status} =
-      List.keyfind(args, :status, 0, {:status, :internal_server_error})
-    {:message, message} =
-      List.keyfind(args, :message, 0, {:message, status})
-    code = Plug.Conn.Status.code(status)
+defmodule PlugRest.ServerError do
+  @moduledoc """
+  Raised for Server errors
+  """
 
-    %PlugRest.ResourceError{message: message, plug_status: code}
+  defexception [:message, plug_status: 500, conn: nil, handler: nil]
+
+  def message(ex) do
+    status = PlugRest.Conn.Status.status(ex.plug_status)
+    "#{status} for #{ex.conn.method} #{ex.conn.request_path} " <>
+      "(#{inspect ex.handler})"
+  end
+end
+
+defmodule PlugRest.RequestError do
+  @moduledoc """
+  Raised for Request errors
+  """
+
+  defexception [:message, plug_status: 400, conn: nil, handler: nil]
+
+  def message(exception) do
+    PlugRest.ServerError.message(exception)
   end
 end