+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
-## v0.9.0-dev
+## v0.9.0
 
 * Enhancements
   * Don't catch exceptions raised by resource callbacks. This allows
   `Plug.Debugger` to render the errors if it's enabled.
 
 * Backwards incompatible changes
+  * For HTTP error codes, your router needs `use Plug.ErrorHandler`
   * The resource `init/2` callback will only terminate when `:stop` is
     returned. All other value will continue REST execution.
   * If the router can't find a match it will raise a `NoRouteError`.
+  * Connections returning 400 or 500 will raise `RequestError` and
+    `ServerError`, respectively.
 
 ## v0.8.0